+++ smarts/core/scenario.py
@@ -939,6 +939,21 @@ class Scenario:
         """The map geometry filepath."""
         return os.path.join(self._root, "map.glb")
 
+    @property
+    def map_glb_metadata(self):
+        """The metadata for the current map glb file."""
+        metadata = self.map_glb_meta_for_file(self.map_glb_filepath)
+        return metadata
+
+    @staticmethod
+    @lru_cache(1)
+    def map_glb_meta_for_file(filepath):
+        """The map metadata given a file."""
+        import trimesh
+
+        scene = trimesh.load(filepath)
+        return scene.metadata
+
     def unique_sumo_log_file(self):
         """A unique logging file for SUMO logging."""
         return os.path.join(self._log_dir, f"sumo-{str(uuid.uuid4())[:8]}")